fix saving file upload

+++ src/controllers/user.ts
@@ -15,15 +15,19 @@ export const update = async (req: Request, res: Response) => {
       userImage: req.file ? "assets/" + req.file.filename + "." + req.file.mimetype.split("/")[1] : req.user.userImage,
     };
 
-    const user = await usersModel.findByIdAndUpdate({ _id: user_id }, data, { new: true });
-
     if (req.file) {
       const imageType = req.file.mimetype.split("/");
       if (imageType[0] !== "image" || req.file.size > 1048576) {
         await fs.remove("src/temp/" + req.file?.filename);
         return res.status(200).json({ code: 200, message: imageType[0] !== "image" ? "failed update, file must image format such as jpeg, jpg, png" : "Size file can't more than 1024kb or 1mb" });
       }
-      if (user) await savefile("assets", req.file?.filename + "." + imageType[1]);
+    }
+
+    const user = await usersModel.findByIdAndUpdate({ _id: user_id }, data, { new: true });
+
+    if (user && req.file) {
+      const imageType = req.file.mimetype.split("/");
+      await savefile("assets", req.file?.filename + "." + imageType[1]);
       await fs.remove(req.user.userImage);
     }
